# added inlining to proc_macros and moved benches to dedicated folder for pnet_packet